fix nginx build file access

+++ install.sh
@@ -1,3 +1,10 @@
+#!/bin/sh
+
+if [ ! -f .env ]; then
+  echo "Error: .env file does not exist"
+  exit 1
+fi
+
 docker builder prune -af
 docker system prune -af
 
@@ -6,7 +13,9 @@ docker compose down -v
 docker compose build --no-cache
 docker compose up -d
 
-sleep 2
-
 chown -R www-data:www-data docker/laravel/
 docker compose exec -T laravel_app php artisan migrate
+
+rm ./docker/laravel/public/build/ -r
+
+docker compose exec -T node npm run build
